Refactored CLI interface
