+++ .babelrc.js
@@ -1,4 +1,7 @@
 module.exports = {
-    presets: [["@babel/preset-env", { targets: { node: "current" } }], "@babel/preset-typescript"],
-    plugins: ["@babel/plugin-transform-spread"]
+  presets: [
+    ["@babel/preset-env", { targets: { node: "current" } }],
+    "@babel/preset-typescript"
+  ],
+  plugins: ["@babel/plugin-transform-spread"]
 };
