+++ .eslintrc.json
@@ -1,0 +1,14 @@
+{
+  "parser": "@typescript-eslint/parser",
+  "extends": ["plugin:@typescript-eslint/recommended"],
+  "parserOptions": { "ecmaVersion": 2021, "sourceType": "module" },
+  "plugins": ["@typescript-eslint"],
+  "env": { "node": true, "es6": true, "mocha": true, "jest": true },
+  "rules": {
+    "no-console": "off",
+    "semi": ["error", "always"],
+    "handle-callback-err": "error",
+    "@typescript-eslint/no-var-requires": "off",
+    "no-constant-condition": "off"
+  }
+}
