--- .prettierrc
@@ -1,11 +1,0 @@
-{
-    "printWidth": 120,
-    "trailingComma": "none",
-    "useTabs": false,
-    "tabWidth": 4,
-    "semi": true,
-    "singleQuote": false,
-    "jsxBracketSameLine": true,
-    "jsxSingleQuote": false,
-    "arrowParens": "avoid"
-}
