+++ src-old/create-full-stack-app.ts
@@ -1,0 +1,99 @@
+import { createApp } from "./create-app";
+import { resolve } from "path";
+import { installDependencies } from "./install-dependencies";
+
+export const createFullStackApp = async (name: string): Promise<void> => {
+  await createApp(name, resolve(__dirname, "../templates/full-stack"), [
+    "dist/client"
+  ]);
+
+  await installDependencies(
+    name,
+    [
+      "axios",
+      "bcryptjs",
+      "connect-redis",
+      "cookie-parser",
+      "csurf",
+      "debug",
+      "dotenv",
+      "ejs",
+      "express",
+      "express-favicon",
+      "express-session",
+      "express-validator",
+      "helmet",
+      "html-to-text",
+      "http-errors",
+      "jsonwebtoken",
+      "moment",
+      "mongodb",
+      "morgan",
+      "multer",
+      "nodemailer",
+      "numeral",
+      "randomstring",
+      "redis",
+      "uuid",
+      "validator"
+    ],
+    [
+      "@babel/core",
+      "@babel/plugin-transform-spread",
+      "@babel/preset-env",
+      "@types/bcryptjs",
+      "@types/connect-redis",
+      "@types/cookie-parser",
+      "@types/csurf",
+      "@types/debug",
+      "@types/express",
+      "@types/express-session",
+      "@types/html-to-text",
+      "@types/jsonwebtoken",
+      "@types/morgan",
+      "@types/multer",
+      "@types/node",
+      "@types/nodemailer",
+      "@types/numeral",
+      "@types/randomstring",
+      "@types/react",
+      "@types/react-dom",
+      "@types/react-router-dom",
+      "@types/redis",
+      "@types/uuid",
+      "@types/validator",
+      "@types/webpack",
+      "@typescript-eslint/eslint-plugin",
+      "@typescript-eslint/parser",
+      "autoprefixer",
+      "babel-loader",
+      "bson-objectid",
+      "concurrently",
+      "css-loader",
+      "eslint",
+      "eslint-config-defaults",
+      "eslint-plugin-react",
+      "faker",
+      "file-loader",
+      "formik",
+      "html-entities",
+      "node-sass",
+      "nodemon",
+      "postcss",
+      "postcss-loader",
+      "prettier",
+      "react",
+      "react-dom",
+      "react-router-dom",
+      "react-simple-widgets",
+      "sass-loader",
+      "style-loader",
+      "ts-loader",
+      "typescript",
+      "url-loader",
+      "webpack",
+      "webpack-cli",
+      "yup"
+    ]
+  );
+};
